requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-agents==1.1.0b3
azure-ai-projects==1.0.0b12
azure-core==1.35.0
azure-identity==1.23.0
azure-storage-blob==12.27.0b1
azure-ai-evaluation==1.9.0
azureopenai==0.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
distro==1.9.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.41
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
openai==1.93.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
setuptools==78.1.1
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1
